Fix error when a dictionary changes while iterating through it
While `await`ing `ws.send()`, if a client got dropped or rejected
a websocket message, the `ws_clients` dict could change in the
middle of an iteration. This causes Python to raise an exception.

Instead, first save a snapshot of clients local to this coroutine as a
list that can't be modified by some other code in the event loop. This
may cause some other exception; the quart docs are unclear on this and
I have not been able to repro.