Use `key` as workaround for immutable name prop
The BasemapLayer component's `name` prop is immutable, so changing it
has no effect on the DOM. We use React's `key` prop, which effectively
reinitializes the component whenever the value passed into this prop
changes. By setting it to the baselayer name, the key changes in exactly
the cases where we need the component to be re-rendered.